requirements.txt
fastapi
pillow
PyPDF2
jinja2
uvicorn
python-multipart
